% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioassay}
\alias{bioassay}
\title{bioassay}
\format{
A data frame with 4 rows and 3 variables:
\describe{
  \item{x}{dose [log g/ml]}
  \item{n}{number of animals}
  \item{y}{number of deaths}
}
}
\usage{
bioassay
}
\description{
Bioassay data set. Taken from Table 3.1
in the BDA book
}
\keyword{datasets}
