% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioassay_posterior}
\alias{bioassay_posterior}
\title{bioassay_posterior}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
  \item{alpha}{Parameter of the bioassay model}
  \item{beta}{Parameter of the bioassay model}
}
}
\usage{
bioassay_posterior
}
\description{
Sample from the posterior of the bioassay model.
}
\keyword{datasets}
