% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassaylp.R
\name{bioassaylp}
\alias{bioassaylp}
\title{Unnormalized log-posterior for bioassay, assuming uniform prior}
\usage{
bioassaylp(a, b, x, y, n)
}
\description{
Unnormalized log-posterior for bioassay, assuming uniform prior
}
