% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvnorm.R
\name{dmvnorm}
\alias{dmvnorm}
\title{Computes the density of a multivariate normal distribution}
\usage{
dmvnorm(x, mean, sigma, log = FALSE)
}
\arguments{
\item{x}{vector or matrix of values for which the density is computed.
If \code{x} is a vector, it represents a single draw.
If \code{x} is a matrix, each row is taken to be a single draw.}

\item{mean}{mean vector}

\item{sigma}{covariance matrix}

\item{log}{logical indicating whether to return on the log scale or not.
False by default.}
}
\value{
Densities of the draws x.
}
\description{
Computes the density of a multivariate normal distribution
}
