% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnew.R
\name{dtnew}
\alias{dtnew}
\title{The density of a transformed t-distribution}
\usage{
dtnew(x, df, mean = 0, scale = 1, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{df}{The number of degrees of freedom}

\item{mean}{Vector of means}

\item{scale}{Vector of scale/standard deviations}
}
\description{
The density of a transformed t-distribution
}
