% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilpisjarvi}
\alias{kilpisjarvi}
\title{kilpisjarvi}
\format{
A data frame with 62 rows and 5 columns:
\describe{
  \item{year}{Year}
  \item{temp.june}{Average temperature in June}
  \item{temp.july}{Average temperature in July}
  \item{temp.august}{Average temperature in August}
  \item{temp.summer}{Average temperature in July--August}
}
}
\usage{
kilpisjarvi
}
\description{
Kilpisjärvi data set. The data contains the Kilpisjärvi
summer month temperatures 1952--2013. Kilpisjärvi is in very northern
part of Finland. Data by Finnish Meteorological Institute (CC-BY 4.0).
Name of the observation station in FMI database is "Enontekiö Kilpisjärvi
kyläkeskus"
}
\keyword{datasets}
