% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassaylp.R
\name{log_inv_logit}
\alias{log_inv_logit}
\title{Implementation of log(1 / (1 + exp(-x))) robust to over- and under-flow}
\usage{
log_inv_logit(x)
}
\arguments{
\item{x}{numeric input}
}
\description{
Implementation of log(1 / (1 + exp(-x))) robust to over- and under-flow
}
