% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcse_quantile.R
\name{mcse_quantile}
\alias{mcse_quantile}
\title{Computes MCSE for quantile estimates based on independent draws}
\usage{
mcse_quantile(draws, prob)
}
\arguments{
\item{draws}{Monte Carlo draws}

\item{prob}{probability for which quantiles is computed}
}
\description{
Computes MCSE for quantile estimates based on independent draws
}
