% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Produces random draws from a multivariate normal distribution}
\usage{
rmvnorm(n, mean, sigma)
}
\arguments{
\item{n}{number of draws}

\item{mean}{mean vector}

\item{sigma}{covariance matrix}
}
\description{
Produces random draws from a multivariate normal distribution
}
