% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnew.R
\name{StudentT}
\alias{StudentT}
\alias{dtnew}
\alias{ptnew}
\alias{qtnew}
\alias{rtnew}
\title{The Student-t Distribution}
\usage{
dtnew(x, df, mean = 0, scale = 1, log = FALSE)

ptnew(q, df, mean = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qtnew(p, df, mean = 0, scale = 1)

rtnew(n, df, mean = 0, scale = 1)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{df}{Vector of degrees of freedom.}

\item{mean}{Vector of location/mean values.}

\item{scale}{Vector of scale values.}

\item{log, log.p}{Logical; If \code{TRUE}, values are returned on the log scale.}

\item{lower.tail}{Logical; If \code{TRUE} (default), return P(X <= x).
Else, return P(X > x) .}

\item{p}{Vector of probabilities.}

\item{n}{Number of samples to draw from the distribution.}
}
\description{
Density, distribution function, quantile function and random generation for
the Student-t distribution with location/mean \code{mean}, scale
\code{scale}, and degrees of freedom \code{df}. See
https://en.wikipedia.org/wiki/Location\%E2\%80\%93scale_family for more details
on how to get from the standard t-distribution to the t-distribution with
additional location and scale parameters.
}
\seealso{
\code{\link[stats:TDist]{TDist}}
}
