% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windshieldy1}
\alias{windshieldy1}
\title{windshieldy1}
\format{
A vector with 9 elements
}
\usage{
windshieldy1
}
\description{
windshieldy1 data set. The data represents a sample of
windshields whose hardness has been measured.
}
\keyword{datasets}
