% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windshieldy2}
\alias{windshieldy2}
\title{windshieldy2}
\format{
A vector with 13 elements
}
\usage{
windshieldy2
}
\description{
windshieldy2 data set. The data represents a sample of
windshields whose hardness has been measured.
}
\keyword{datasets}
