% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioassay_posterior_density_plot.R
\name{bioassay_posterior_density_plot}
\alias{bioassay_posterior_density_plot}
\title{Plot the posterior density in a grid using the implemented function p_log_posterior}
\usage{
bioassay_posterior_density_plot(alpha_limits = c(0, 1),
  beta_limits = c(0, 10))
}
\description{
Plot the posterior density in a grid using the implemented function p_log_posterior
}
